ocr_fastapi/requirements.txt
pydantic
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
dspy
python-dotenv

